move parse_cfgspecs to rustc_interface
